+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * !! Major breaking change !! Deprecating "ADVERTISING_ID" as device ID generation strategy. SDK will fall back to 'OPEN_UDID'. All "ADVERTISING_ID" device ID's will have their type changed to "OPEN_UDID". If the device will have a "null" device ID, a random one will be generated.
 * !! Major breaking change !! Changing device ID without merging will now clear all consent. It has to be given again after this operation.
 * !! Major breaking change !! Entering temporary ID mode will now clear all consent. It has to be given again after this operation.
+* User profile issue fixed, user data was deleted for key if no value, null or undefined value was provided against that key.
 * Device ID can now be changed when no consent is given
 * Push notification now display/use the sent badge number in Android. It's visualization depends on the launcher.
 * When recording internal events with 'recordEvent', the respective feature consent will now be checked instead of the 'events' consent.